+++ java-io-http/src/main/java/internal/io/http/ext/CachedHttpResponse.java
@@ -24,9 +24,7 @@ public final class CachedHttpResponse implements HttpResponse {
     @Override
     public @NonNull MediaType getContentType() throws IOException {
         String contentTypeOrNull = cached.getHeaders().firstValue(HttpHeaders.HTTP_CONTENT_TYPE_HEADER).orElse(null);
-        if (contentTypeOrNull == null) {
-            throw new IOException("Missing content-type in HTTP response header");
-        }
+        if (contentTypeOrNull == null) return NO_CONTENT_TYPE;
         try {
             return MediaType.parse(contentTypeOrNull);
         } catch (IllegalArgumentException ex) {