+++ java-io-http/src/main/java/nbbrd/io/http/ext/CachedResponse.java
@@ -23,11 +23,6 @@ public class CachedResponse {
     int statusCode;
 
     /**
-     * HTTP reason phrase.
-     */
-    String reasonPhrase;
-
-    /**
      * Response headers.
      */
     @NonNull