+++ java-io-http/src/test/java/nbbrd/io/http/HttpClientTest.java
@@ -29,6 +29,7 @@ import org.junit.jupiter.api.extension.RegisterExtension;
 import javax.net.ssl.*;
 import java.io.IOException;
 import java.io.InputStream;
+import java.net.ProxySelector;
 import java.net.URI;
 import java.net.UnknownHostException;
 import java.nio.charset.StandardCharsets;
@@ -347,6 +348,37 @@ public abstract class HttpClientTest {
         }
 
         wire.verify(1, getRequestedFor(urlEqualTo("/abc/../first.xml")));
+    }
+
+    @Test
+    public void testContentType() throws IOException {
+        HttpContext context = HttpContext
+                .builder()
+                .proxySelector(ProxySelector::getDefault)
+                .sslSocketFactory(this::wireSSLSocketFactory)
+                .hostnameVerifier(this::wireHostnameVerifier)
+                .build();
+        HttpClient x = getClient(context);
+
+        wire.resetAll();
+        wire.stubFor(get(SAMPLE_URL).willReturn(ok()));
+
+        try (HttpResponse response = x.send(HttpRequest.builder().query(wireURL(SAMPLE_URL)).headers(GENERIC_DATA_21_HEADER).build())) {
+            assertThat(response.getContentType()).isEqualTo(HttpResponse.NO_CONTENT_TYPE);
+        }
+
+        wire.verify(1, getRequestedFor(urlEqualTo(SAMPLE_URL)));
+
+        wire.resetAll();
+        wire.stubFor(get(SAMPLE_URL).willReturn(okForContentType("/ / /", "body")));
+
+        try (HttpResponse response = x.send(HttpRequest.builder().query(wireURL(SAMPLE_URL)).headers(GENERIC_DATA_21_HEADER).build())) {
+            assertThatIOException()
+                    .isThrownBy(response::getContentType)
+                    .withMessageContaining("Invalid content-type in HTTP response header");
+        }
+
+        wire.verify(1, getRequestedFor(urlEqualTo(SAMPLE_URL)));
     }
 
     @Test